Move QuotasService to services package
